ipq807x: backport MHI/QRTR start patches
In kernel 5.11 MHI core has dropped the auto start
feature that it used previously, and thus removed it
from ath11k as well.
This obviously wasn't back-ported to 5.10 kernel, so it
causes a big issue for our use of backports newer than
5.10 as ath11k will put the MHI cards in POWER_ON
state and then just wait forever.
Issue is that nothing actually starts the MHI channels,
so instead of hacking ath11k to revert this lets
backport MHI removel of auto-start and then QRTR
patches that will start and remove channels.

Signed-off-by: Robert Marko <robimarko@gmail.com>